Delay warning of base class changes until necessary.
The version/checksum stored in the TStreamerBase is used only during member wise streaming and thus
we must only warn if the class is read as part of a TClonesArray or memberwise streamed collection.
In all the other cases, the discrepancy does not matter as the information is not used (the version
information is part of the byte stream itself).