Restore R5 typedef behaviour in presence of LinkDef file
this is necessary since in R5 the typedef pragma selected class
is present to allow cint to know about these typedefs.
This is not necessary for Root6.
On the other hand, we need autoload entries for these typedefs.
The solution is to keep them in the list of selected typedefs but
not to "drill a hole" through them and select the class they point
to.
This behaviour is present only for genreflex + selection xml, for
which the typedefs are "followed up": the pointed class is selected
via the typedef. Again, a behaviour of ROOT5 preserved in ROOT6.