Reduce the number of Flush (fsync) when closing a TFile.
Rather than doing one call to Flush (fsync) per TDirectory during the execution of TFile::Close
we reduce this to just one call.  Doing so many sync was extremely slow on an ATA disk (order
of 5 minutes for 90000 TDirectories.  With this update the time is down to half a second.