Array size attribute must be at start of data member comment. Update CheckSum calculation.
The array information must (as was always intended but never inforced) but at
the beginning of the comment (only preceded by the comment start and white
spaces):

   //[fArraySize] array of size fArraySize

but not:

   // my comment [some details] etc.

Implement this search in TVirtualStreamerInfo::GetElementCounterStart

Add kReflexV5 and kNoRangeCheck as CheckSum alternative.

CheckSum alternative calculation newer than kNoRangeCheck will only use information in
between brackets that is at the start of the comment.

kReflexV5 emulates the way the checksum was calculated when the dictionary
was generated with genreflex (no typedef and no comment information).