Don't put in rootmaps headers for templates
as this is not well defined.
For example, myVector<T> defined in myVector.h and classes  A, B
defined elsewhere.
The rootmap entry for myVector<A> and myVector<B> is going to be
myVector.h and if myVector<A> and myVector<B> reside in 2 libraries
we will have the same entry for 2 libraries.