Do not split classes with a custom streamer.
Classes for which a Streamer function was externally provided are no longer
split; they were split in v5 if the dictionary was generated via rootcint but
were not split if the dictionary was generated via genreflex.

Classes with a custom Streamer function and classes with an older, non StreamerInfo
based automatic streamer are also no longer split.

To force the splitting of those classes, thus by-passing the custom Streamer,
when storing the object in a TTree and in a collection object, use: