Enable header parsing on demand for functions
Steps accomplished in this commit:
1) Selection rules: enable the "pragma link function" and related
matching
2) Scanner: take appropriate action when visiting a FunctionDecl
3) RootCling: associate headers to FunctionDecls and store it in the
catalogue dedicated to class2header (now decl2header) association.

It is worth noting that it would be sufficient to store the function
names in the rootmaps to enable (for free!) "autoload on functions".
Indeed the same mechanisms of hpod and autoload of classes work for
functions.