Touch libCore_rdict.pcm as part of the dictionary creation.
The rootmap files' target is only used when invoked explicitly, else they are generated as a side
effect of generating the dictionaries. Thus the rootmap rule for libCore is never called during a
regular ROOT build, and its PCM must instead be "created" during the creation of G__Core.cxx.

This fixes a missing prerequisite / continuous rebuild of the dictionaries.