TStreamerElement now records the normalized type name
Rather than record the name with potentially some typedefs, we now record
the normalized name (stripping all but the 'opaque' typedefs).  With the
typedef in the name, the ROOT file is not self describing ; adding the typedef
information to the ROOT file would make it self describing but would introduce
another layer of complication when having to deal with typedefs that changed
from one type to the other over time.