Update CheckSum Reflex like calculation when dealing with enums.
When a classes went through genreflex and Cintex any data member that
had an enum type was seens as having the type 'int', in particular for
the CheckSum calculation.  So in the calculation of the Reflex style
CheckSum we now replace all enum type by int.  This addresses part of ROOT-6271.