Fix ROOT-4549 (TRefArray missing cleanup).
TRefArray does not clean the fUIDs array during TRefArray::Streamer read in. This means that when a TRefArray is serialized in, e.g. from a tree, it can have old data still in the fUID array. This manifests when Compress() is called because then the old data get recognized in the array (fLast is updated), which then enables TRefArray to accept invalid accesses.