Add TClassEdit::IsSTLCont 'specicalization'.
In the default case of IsSTLCont (where we want to know the container
type and do not care about which allocator is used), we do not need
to do a full parsing of the class name, we just have to pick at the
beginning.  This significantly speeds up the call.