Delayed parsing: Treat case where d.member is of type of non selctd intern. cl
There are classes like
class VeloAlignCond{
public:
   struct PositionPaths {
      int i;
      };
private:
   PositionPaths a;
};
in LHCb and the selection only skims the VeloAlignCond class. This is done probably
for interactivity only classes.
See for example http://lhcb-release-area.web.cern.ch/LHCb-release-area/DOC/brunel/latest_doxygen/d2/d09/_velo_align_cond_8h_source.html
If delayed parsing is active, some steps have to be taken in order to properly
treat such cases:
o When filling the TClass entry from the protoclass information, if a
TClass entry for the internal class, PositionPaths in this case, is not there
a new entry is created with the state kForwardDeclared.
o Such operation is now possible and does not trigger any autoparsing.
o As in ROOT5, if afterwards one tries to do IO with a class like the above one
without selecting the internal class, a warning message is prompted
"Warning in <TStreamerInfo::Build>: VeloAlignCond: VeloAlignCond::PositionPaths
has no streamer or dictionary, data member "a" will not be saved"