Remove redundant portion of code generated for dictionaries
as now the protoclasses carry the information.
Since we do not persist the ast in form of pcms (clang pcms), we
were obliged to transfer some annotations of the decls present
at dictionary generation time in the AST to the AST at runtime
(and of course, ultimately TClass) using c++ code in the
dictionaries.
Now, with ROOT pcms and protoclasses, the information is passed w/o
the need of these constructions.