Introduce fCompOpt and fCompFull in TStreamerInfo
fCompOpt and fCompFull holds pointer to the TCompInfo
respectively for the set of optimized elements (and the
non-optimizable elements) and the full set of elements
(non optimized).

The set of optimized elements (and the non-optimizable elements)
are held are the start of fComp.  The set of original TCompInfo
corresponding to the optimized ones is held at the end of fComp.

This is the first step towards always keeping both optimized
and non-optimized elements to avoid the need for re-compilation.