Adapt to the code to use the intended (list of) StreamerElement
This is ‘instead’ on re-compiling the StreamerInfo to make the optimized
list to be the list (optimized/non-optimized) the calling code needed.

Add to TVirtualStreamerInfo the routines:

    TStreamerElement *GetElement(Int_t id) const;
    Int_t  GetElementOffset(Int_t id) const;

respectively  as a convenience accessor and as a performance accessor
to the StreamerElement information based on ‘id’ being the index
in the full list of StreamerElement (in contrast GetElem and GetOffset
take an id is the index in the shorted list of optimized elements).

Add TStreamerInfo::GetNelement() as a short cut to  fElements->GetEntries();