Support for XRootD 4
This patch - together with  cc884ecb832f025af4dd5eea0eb27ad017e76b0d - adds support fro XRootD version 4.
The bulk of the patch copes with the changes in the XRootD network classes to support IPv6. The rest is mostly
handling of the changes in packaging.
The PROOF part still uses the old XrdClient (still delivered with XRootD).