Enable autoloading even without a TApplication.
In commit 73449447, the enabling of the autoloading was moved from TROOT to
TApplication to delay it until after main starts (i.e. "after TROOT has been
fully created and all linked libraries have been initialized.").

This had the somewhat unintentional side-effect of disabling the autoloading
if the user's executable did not have a TApplication.

Now that the initialization of the interpreter must also be delayed after the
start of main (see commit 73449447 and de8922d and TROOT.cxx comments) we
can move the enabling of the autoloading back to TROOT (but this time in
TROOT::InitInterpreter rather than the constructor).

Also prevent any enable of the interpreter autoloading when library loading is
not allowed.