Fix ROOT-6592, TClassEdit name normalization handling of param which are addr.
TClassEdit::GetNormalizedName was mis-handling:

    tempname<&ComplexElement::level>

[in this case ComplexElement::level is a function].  The issue that
GetSplit was adding a spot for the trailing pointers/stars (and references, etc.)
only if present.  Consequently the calling code had to 'guess' whether the
last slot in list was the last parameters or the trailing stars.

To resolve this, GetSplit now always add the slot for this trailing information
and the caller needs to properly use it or discard it.