Make TMVA thread-safe with respect to use of Reader
It is now possible to create independent TMVA::Readers and use
them simultaneously on different threads.
Training of MVAs is still only safe single-threaded. In addition,
it is not safe to use multiple instances of MethodCFMlpANN either
single or multi-threaded because of a global 'this' pointer.

Danilo:
Properly take care of the deletion of some atomic pointers

Signed-off-by: Danilo Piparo <danilo.piparo@cern.ch>