Support PyROOT's naming convention in name normalization.
For the shadow of python classes in ROOT, the name begin passes include the module name separated
from the class name via a dot (i.e. follow python convention and use something other than :: to be
able to distinguish (for now) python module and C++ namespace ... this difference should probably
be removed.