Add iterator interface for TTreeReader.
We want sequential reads, thus use a forward iterator.
The iterator does not really reference any data but the tree entry number
so return that as the deref-op. The same operator (also called from the
increment operators) sets its entry number in the associated TTreeReader.

This iterator is "C++14 safe" in that its past-end-iterator can be value
constructed and all past-end-iterators compare equal.