TClassEdit once again strips inline namespace ... but only when known.
Since we rely much more on TClassEdit to normalize the name (in part to avoid
autoparsing), we need to strip the inline namespace whenever possible.

This feature had been remove by commit 4480ef431445de8383e27b828a1e907f0e03f150.
We reintroduce it we a twist where rather than guessing by looking at the name,
we look for the scope in clang type database and ask whether the scope
exist and is inline.  We limit this search to immediate sub scope of std and
which are not templated, we also cache the result.