Complete rewrite of TClassEdit::ResolveTypedef.
Properly resolve all typedefs including those used a part of the scope.

The implementation tries to reduce the number of memory copy (in cases where
there is no typedef, there should not be any copy or string creation ... this is
not yet the case in all cases as the underlying routines (GetPartiallyDesugaredType
for example) sometimes returns a 'new'/'different' type eventhough there is
no effective differences.  This implementations deals better with nested
typedef, mix of const, pointer and references and is more consistent with
spaces.