Disable the loading of the I/O rules form the file.
The problem stems from the (current dormant/unused) support for the rules in the
ROOT File. When loaded they do not seem to be reduplicated and hence accumulated
and need to be verified against this growing number. Simply disabling the loading
(as it is not useful at the moment anyway) should solve the issue.