Fix handling of IgnoreTObjectStreamer.
This fixes ROOT-6645.

Now that we have 2 lists of compiled information fCompOpt and fCompFull
and that fCompFull is supposed to be containing *all* elements, we can not
just skip the ignore base class.  Instead we put them in fCompFull and out
of fCompOpt.  Consequently we also need to ignore them properly in the
old fashion Read/WriteBuffer (they are not put in the sequence of actions).