ROOT PCM files are intended to be always rebuild, so no need for schema evolution.
By skipping the storing of the StreamerInfo for the ROOT PCM files themselves,
as they are guaranteed to be indentical to the 'loaded' ones, we save both
space (on disk and in memory) and time (skipping the writting and reading).

Note that using the string '?filetype=pcm' added to the filename sets the
file in a mode that skipping writing and reading the StreamerInfo.