Do not check StreamerInfor when reading rootpcms
This is a performance optimisation which aims to a faster startup of root.
Some ~13-15% of the startup time of root is spent checking the StreamerInfo
of rootpcms. While this is fundamental to guarantee portability of rootfiles
despite changes in the classes layout, for rootpcms, by construction, we
have the guarantee that they are built with the same root version which is
reading them.