Full TROOT::EndOfProcessCleanups does too much from PyROOT.
In v5 this includes destroying the global variable and hence
some of the state in CINT, we need to delay this some more.
In v6 (beside that the global variable cleanups is not
implemented yet), this will be fine as the cleanup of variable
will leaves the handle PyROOT might have in a valid state.