Do not expose umbrella to FileManager. Fixed fwd decl callback.
Instead of #including the ubmrella header we now declare its content.
This hides the umbrella header in the include chain, allowing the forward
declarator of cling to detect that the include for TObject is TObject.h
and not the umbrella header. This fixes the "healing" for the forward
declarations when seeing the defining header being included.