Fix for ROOT-6637: NO_SELF_AUTOSELECT functionality is in root6
but implemented using traits.
If a template is present in the selection namespace and it inherits
from the class SelectNoInstance none of its instantiations will be
selected. On the other hand all of the instantiations selected with
other mechanisms will inherit the properties specified in the template,
for example transient data members or template arguments to keep.