Implement selection of typedefs with selection xmls
this feature restore the symmetry with Linkdefs and can be leveraged
to limit header parsing making names of typedefs known to the system.
Therefore:
1) If a name of a typedef is used to select the underlying type (e.g.
typedef A myTDforA; <class name="myTDforA" />) the underlying type is
selected and the typedef stored in the rootpcm. This was true before
this commit as well.
2) If a typedef name is selected as <class typedef="myTDforA" />, the
underlying type is not automatically selected but the typedef stored
in the rootpcm.