Check for the proxy w/o relying on name
which makes the check even more powerful: we now do not check for
mother classes not only for stl collections but also for all proxied
collections.
This is perfectly consistent with the concept of collection proxy,
which is there exactly to abstract from the details of the container
and focus on the contained entities.