Add variadic template function CallFunc_SetArguments to properly dispatch args.
template <typename... T>
void CallFunc_SetArguments(CallFunc_t *func, const T&... args)

and is used in place where the elipsis might have been used to dispatch
the arguments to calls to CallFunc_SetArg with the 'right' type. (i.e. in
a typesafe manner).

Also to have the proper conversion add CallFunc_SetArg for char, short, int
to correctly cast them (as it was done in the past) to long and (new with
this patch) to unsigned long (depending of the signedness).