Avoid protoClasses for classes in libCore and libRIO's dictionary.
Since we do not load in rootcling the pch or the header for libCore and libRIO, the
TClass object coming from their dictionary are actually partial and
should not be used. [The TClass used of the I/O of the rootpcm are
complete as we explicitly load their header in rootcling).