Move CheckABICompatibility() to IncrParser; move funcs into unnamed namespace.
The IncrParser has the ability to check the stdlib that the interpreter sees,
versus what the compiler sees. Of course what we want to check is the
interpreter's stdlib - including any -I passed at the prompt. The other
advantage of this change is that system headers are now pulled from the PCH
if it exists (ROOT-6794)

(cherry picked from commit 09185c05a4ffb38b5f2a0dea7218d9c4137e5198)