Simplify the TClass::HasDictionary() method
This method, which is not static, does not need to look for the
dictionary of the current TClass instance in the list of
dictionaries. It is sufficient, by construction, to check that the
instance comes from a library.
This avoid a search in the loaded dictionaries and a normalisation
procedure for the name of the class, which could cause parsing.