Revert "Prevent parsing when creating the collection proxy element"
This reverts commit 5984225a9fed03b0f10037873c9000f9b050b1f4.
This commit aimed to get rid of autoparsing in the TGenCollectionProxy::Value
ctor when the value is a pair: this change was intended to be a performance
optimisation.
Simply turning off autoparsing before interrogating the typesystem about
the element does not have the desired effect: many cmssw tests fail because
of TStreamerInfoElements being nullptr.
A more comprehensive solution to avoid parsing in presence of pairs is to
be identified.