Add autoloading of selected enums
is an enum is selected, it becomes an autoload key. The mechanism
works for queries to the typesystem via TEnum::GetEnum and not for
the interpreter.
In this latter case, presently, the limitation is the absence of the
concept of enum fwd declaration. We cannot hook on the incomplete type
error clang would emit.