[modules] Add a ROOT module map, enabling ROOT to be built with -fmodules
C++ Modules (http://clang.llvm.org/docs/Modules.html) allow significant speed ups
of the compilation times. They will be available with clang 3.6.0.

ROOT will use these modules as a replacement for its dictionaries. We refer to
that feature as runtime C++ modules. They are will be the same as clang's C++
modules, but will be available through cling.

As a first step we need to make standalone clang and LLVM compile with modules
enabled (-fmodules). This is already completed. Secondly we need to build ROOT
with -fmodules. This modulemap is the driver building ROOT with modules. Currently
32 out of 90 libraries can be built.

In order to test it out one needs to copy the module.modulemap in $ROOTSYS/include,
and add to his MyConfig.mk the following:
CXXFLAGS+= -Xclang -fmodules -Xclang -fmodules-cache-path=$(ROOTSYS)/pcm/
CFLAGS+= -Xclang -fmodules -Xclang -fmodules-cache-path=$(ROOTSYS)/pcm/

and of course ROOT needs to be configured to be built with newest clang compiler.

Once ROOT fully compiles with modules, the implementation of runtime modules will
be far easier.