Autoparse also typedefs (ROOT-6883).
The typedef resolution uses cling::LookupHelper::findType() which does
not know whether it should look for a class or a typedef. In above issue,
the type name is actually the name of an explicit template specialization.
This explicit template specialization has an autoparse entry and is thus
available - in principle. But here, autoparsing was explicitly disabled.
Failure to autoparse leads to the type's instantiation from the template
pattern which is fatal.

(cherry picked from commit 33707c145f09acd793d391c3e28314a06c772200)