Adapt the test to the new "filter" imposed to the ListOfTypes
in order to limit the number of lookups into cling, now ROOT interrogates
the lists of enums and (proto)classes before asking the interpreter
if a certain name is a typedef. See
http://root.cern.ch/gitweb?p=root.git;a=commit;h=aa5c57e63859c57efb87e3ae9c130ce71c55162c
as a result, "string" is not anymore a typedef for the typesystem.
Indeed it has its own entry as a class. This is probably more consistent
than in the past, since string is "opaque": we do not want to care about
the implementation details. For example, we do not want to know if
string is an alias for basic_string<char>: this could be (and is) platform
dependent.
For this reason, now "string" has been removed from the list of typedefs
expected by the system.