Do not autoparse the class, but #include headers from libs.
An #include of "TH1F.h" was converted into an autoparse of TH1F (for headers
starting with 'T' and ending on ".h"). This was often doing the right thing; a
better solution (implemented here) would be to #include that header with
autoparsing disabled (which is the crucial part!) to prevent recusrive
autoparsing - but to do that only for those headers that are guaranteed to be
compilable because they "come from a library".

ROOT headers will not be autoparsed (because they are in the PCH), all others
will be autoparsed. For the former, the explicit #include during tryAutoParse()
will prevent recursive autoparsing.

(cherry picked from commit c1a572e966a8cb13286b1425d4ccf49308830bae)