Pass the FileEntry for #inclusions, use that to #include.
Fixes cases where the file is found relative to the include's directory, e.g.
.L /foo/bar/test.C
which has
This header cannot be found with any -I but only relative to the includer.
As we know the full path of the file to be included we will just use that.

Fixes out-of-source roottest/root/meta/autoloading/execAutoloadOnInclusion.C