Perf opt: avoid to parse empty fwd decl payload for base libs
libthread, libCore, libRIO.
rootcling was generating dictionaries with an empty string instead of
a nullptr for the fwd declarations payload. As a result, even if nothing
was there to parse, the overhead of compiling an empty string was involved
(sigh). The effect is not really on the runtime of complex applications
but more on root startup which is now 5% faster.