Add a warning in presence of the conditions described here ROOT-6947
print a warning when a class is selected twice, for example with
its name and a typedef to it.
It does not make sense to select twice the same entity within the
same selection and genreflex helps the user to figure out that
the situation can be improved.
In some sense this is an extension of the mechanism of duplicates
detection with rootmaps.