Install custom assert.h in clang resource dir.
At least on Linux, the system assert.h has no include guards. This causes
re-parsing of it, even if it's included in a PCH. This in turn triggers
errors when running ROOT on a system with a different assert.h.

By adding a custom assert.h with include guards, the parsing is intercepted and
the assert.h as included in the PCH is taken. For the interpreter that's good
enough.

Still missing the CMake version.